Move TestSetupAndValidate to fix broken build
